Some refactorings, improved scenarios
