+++ k6/helpers.js
@@ -1,0 +1,6 @@
+export const getRequestParams = (accessToken) => ({
+  headers: {
+    Authorization: `Bearer ${accessToken}`,
+    "Content-Type": "application/json",
+  },
+});
